Finally got constraints working better. Had to go with a container view. NSLayoutConstraints does not work with transform at all.
Still have not gotten to Data Controller for DetailSettingsTVC. I will have to learn how to use textFields in the prototype cells. My biggest concern is the Number Pad keyboard does not have a return button on it. Right now I setting the targets for this keyBoard in the UITableViewDataSource methods. I also do not want the Data controller to be the delegate for the textfields.